requirements.txt
python-telegram-bot>=20.0
asyncpg>=0.26
pydantic>=1.10
python-dotenv>=0.21
pytest>=7.0
